Allow `projects` to be timed as well ask `tasks` (#13)
The Problem:
============
I have projects that are broken down to checklist-steps that can take
less than a minute to complete. For these I prefer to set a timer on the
`project` rather than the individual task.

OmniToggl can't currently do this because

a) The `validate` function only activates if a single `task` is
selected.
b) The downstream functions expect a `task` as input. Forcing a run with
a `project` selected throws an error.

The Solution:
=============

a) Update `validate` to be `true` for a single `task` or `project`
b) Update the flow to work with either a `task` or `project` as the
trigger.

Co-authored-by: Gavin Montague <gavin.montague@itison.com>